Add test for getting hostname patterns
Auditors: @bbondy
