+++ app/httpsEverywhere.js
@@ -7,6 +7,7 @@ const urlParse = require('url').parse
 const DataFile = require('./dataFile')
 const Filtering = require('./filtering')
 const LRUCache = require('lru_cache/core').LRUCache
+const getHostnamePatterns = require('../js/lib/urlutil').getHostnamePatterns
 
 // Map of ruleset ID to ruleset content
 var db = null
@@ -61,34 +62,6 @@ function getRewrittenUrl (url) {
     }
     return undefined
   }
-}
-
-/**
- * Gets applicable hostname patterns for a given URL. Ex: for x.y.google.com,
- * rulesets matching x.y.google.com, *.y.google.com, and *.google.com are
- * applicable.
- * @param {string} url The url to get hostname patterns for
- * @return {Array.<string>}
- */
-function getHostnamePatterns (url) {
-  var host = urlParse(url).hostname
-  var hostPatterns = [host]
-  var segmented = host.split('.')
-
-  // Since targets can contain a single wildcard, replace each label of the
-  // hostname with "*" in turn.
-  segmented.forEach((label, index) => {
-    // copy the original array
-    var tmp = segmented.slice()
-    tmp[index] = '*'
-    hostPatterns.push(tmp.join('.'))
-  })
-  // Now eat away from the left with * so that for x.y.z.google.com we also
-  // check *.z.google.com and *.google.com.
-  for (var i = 2; i <= segmented.length - 2; ++i) {
-    hostPatterns.push('*.' + segmented.slice(i, segmented.length).join('.'))
-  }
-  return hostPatterns
 }
 
 /**
